+++ Src/Metrics/Core/MetricsBuilder.cs
@@ -7,7 +7,7 @@ namespace Metrics.Core
 {
     public interface MetricsBuilder
     {
-        MetricValueProvider<double> BuildePerformanceCounter(string name, Unit unit, string counterCategory, string counterName, string counterInstance);
+        MetricValueProvider<double> BuildPerformanceCounter(string name, Unit unit, string counterCategory, string counterName, string counterInstance);
         MetricValueProvider<double> BuildGauge(string name, Unit unit, Func<double> valueProvider);
         CounterImplementation BuildCounter(string name, Unit unit);
         MeterImplementation BuildMeter(string name, Unit unit, TimeUnit rateUnit);