+++ Src/Metrics/PerfCounters/PerformanceCounters.cs
@@ -2,13 +2,13 @@
 using System;
 using System.Diagnostics;
 using System.Linq;
+using System.Security.Principal;
 using Metrics.Core;
 using Metrics.Logging;
-using Metrics.MetricData;
 
 namespace Metrics.PerfCounters
 {
-    public static class PerformanceCounters
+    internal static class PerformanceCounters
     {
         private static readonly ILog log = LogProvider.GetCurrentClassLogger();
 
@@ -20,7 +20,7 @@ namespace Metrics.PerfCounters
         private const string LocksAndThreads = ".NET CLR LocksAndThreads";
         private const string Networking = ".NET CLR Networking 4.0.0.0";
 
-        public static void RegisterSystemCounters(MetricsContext context)
+        internal static void RegisterSystemCounters(MetricsContext context)
         {
             context.Register("Available RAM", Unit.MegaBytes, "Memory", "Available MBytes", tags: "memory");
             context.Register("Free System Page Table Entries", Unit.Custom("entries"), "Memory", "Free System Page Table Entries", tags: "memory");
@@ -49,7 +49,7 @@ namespace Metrics.PerfCounters
             context.Register("Physical Disk Reads/sec", Unit.Custom("kb/s"), "PhysicalDisk", "Disk Writes/sec", TotalInstance, f => f / 1024.0, tags: "disk");
         }
 
-        public static void RegisterAppCounters(MetricsContext context)
+        internal static void RegisterAppCounters(MetricsContext context)
         {
             var app = Process.GetCurrentProcess().ProcessName;
 
@@ -90,7 +90,44 @@ namespace Metrics.PerfCounters
             Func<double, double> derivate = null,
             MetricTags tags = default(MetricTags))
         {
-            log.Debug(() => string.Format("Registering performance counter Name {0} Category {1} Instance {2}", counter, category, instance));
+            try
+            {
+                WrappedRegister(context, name, unit, category, counter, instance, derivate, tags);
+            }
+            catch (UnauthorizedAccessException x)
+            {
+                var message = "Error reading performance counter data. The application is currently running as user " + GetIdentity() +
+                   ". Make sure the user has access to the performance counters. The user needs to be either Admin or belong to Performance Monitor user group.";
+                MetricsErrorHandler.Handle(x, message);
+            }
+            catch (Exception x)
+            {
+                var message = "Error reading performance counter data. The application is currently running as user " + GetIdentity() +
+                   ". Make sure the user has access to the performance counters. The user needs to be either Admin or belong to Performance Monitor user group.";
+                MetricsErrorHandler.Handle(x, message);
+            }
+
+            return context;
+        }
+
+        private static string GetIdentity()
+        {
+            try
+            {
+                return WindowsIdentity.GetCurrent().Name;
+            }
+            catch (Exception x)
+            {
+                return "[Unknown user | " + x.Message + " ]";
+            }
+        }
+
+        private static void WrappedRegister(MetricsContext context, string name, Unit unit,
+            string category, string counter, string instance = null,
+            Func<double, double> derivate = null,
+            MetricTags tags = default(MetricTags))
+        {
+            log.Debug(() => string.Format("Registering performance counter [{0}] in category [{1}] for instance [{2}]", counter, category, instance ?? "none"));
 
             if (PerformanceCounterCategory.Exists(category))
             {
@@ -107,13 +144,12 @@ namespace Metrics.PerfCounters
                         {
                             context.Advanced.Gauge(name, () => new DerivedGauge(new PerformanceCounterGauge(category, counter, instance), derivate), unit, counterTags);
                         }
-                        return context;
+                        return;
                     }
                 }
             }
 
-            log.ErrorFormat("Performance counter does not exist Name {0} Category {1} Instance {2}", counter, category, instance);
-            return context;
+            log.ErrorFormat("Performance counter does not exist [{0}] in category [{1}] for instance [{2}]", counter, category, instance ?? "none");
         }
     }
 }