+++ Src/Metrics/RemoteMetrics/RemoteMetricsContext.cs
@@ -45,9 +45,12 @@ namespace Metrics.RemoteMetrics
         public override MetricsDataProvider DataProvider { get { return this; } }
         public MetricsData CurrentMetricsData { get { return this.currentData; } }
 
-        public override void Dispose(bool disposing)
+        protected override void Dispose(bool disposing)
         {
-            using (this.scheduler) { }
+            if (disposing)
+            {
+                using (this.scheduler) { }
+            }
             base.Dispose(disposing);
         }
     }