+++ Src/Metrics/Reporters/AsyncScheduledReporter.cs
@@ -1,23 +1,24 @@
 ﻿using System;
 using System.Threading;
+using System.Threading.Tasks;
 using Metrics.MetricData;
 using Metrics.Utils;
 
 namespace Metrics.Reporters
 {
-    public sealed class ScheduledReporter : IDisposable
+    public sealed class AsyncScheduledReporter : IDisposable
     {
         private readonly Scheduler scheduler;
         private readonly TimeSpan interval;
 
-        private readonly Func<MetricsReporter> reporter;
+        private readonly Func<AsyncMetricsReporter> reporter;
         private readonly MetricsDataProvider metricsDataProvider;
         private readonly Func<HealthStatus> healthStatus;
 
-        public ScheduledReporter(Func<MetricsReporter> reporter, MetricsDataProvider metricsDataProvider, Func<HealthStatus> healthStatus, TimeSpan interval)
-            : this(reporter, metricsDataProvider, healthStatus, interval, new ActionScheduler()) { }
+        public AsyncScheduledReporter(string name, Func<AsyncMetricsReporter> reporter, MetricsDataProvider metricsDataProvider, Func<HealthStatus> healthStatus, TimeSpan interval)
+            : this(name, reporter, metricsDataProvider, healthStatus, interval, new ActionScheduler()) { }
 
-        public ScheduledReporter(Func<MetricsReporter> reporter, MetricsDataProvider metricsDataProvider, Func<HealthStatus> healthStatus, TimeSpan interval, Scheduler scheduler)
+        public AsyncScheduledReporter(string name, Func<AsyncMetricsReporter> reporter, MetricsDataProvider metricsDataProvider, Func<HealthStatus> healthStatus, TimeSpan interval, Scheduler scheduler)
         {
             this.reporter = reporter;
             this.metricsDataProvider = metricsDataProvider;
@@ -26,9 +27,9 @@ namespace Metrics.Reporters
             this.scheduler = scheduler;
         }
 
-        private void RunReport(CancellationToken token)
+        private Task RunReport(CancellationToken token)
         {
-            reporter().RunReport(this.metricsDataProvider.CurrentMetricsData, this.healthStatus, token);
+            return reporter().RunReportAsync(this.metricsDataProvider.CurrentMetricsData, this.healthStatus, token);
         }
 
         public void Start()