+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-###0.2.7 / 2014-10-17
+###0.2.10 / 2014-11-06
+* fix error logging for not found performance counters
+
+###0.2.9 / 2014-11-05
+* record active sessions for timers
+
+###0.2.8 / 2014-10-29
+* handle access issues for perf counters
+
+###0.2.7 / 2014-10-28
 * preparations for out-of-process metrics
 
 ###0.2.6 / 2014-10-17